+++ source/leetcode/return-length-of-arguments-passed/return-length-of-arguments-passed-variadic-function.js
